Began simplifying api
