+++ package.json
@@ -8,6 +8,7 @@
   },
   "scripts": {
     "start": "react-scripts start",
+    "server": "node index.js",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject",
